+++ Pipfile
@@ -1,0 +1,14 @@
+[[source]]
+url = "https://pypi.org/simple"
+verify_ssl = true
+name = "pypi"
+
+[packages]
+requests = "*"
+pygithub = "*"
+argparse = "*"
+
+[dev-packages]
+
+[requires]
+python_version = "3.9"
