+++ README.md
@@ -3,3 +3,18 @@
 This secret scanner aims to have no false positives.
 Plus, there's more!  Stay tuned....
 
+## Local Development
+**Prerequisites:**
+* Have Python 3.9 installed
+* Have pipenv installed (`pip3 install pipenv`)
+
+**Steps:**
+* Setup virtualenv and install dependencies
+```
+pipenv install
+```
+
+* Shell into virtualenv
+```
+pipenv shell
+```
